declarative websockets upd
